@@ -2,16 +2,17 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Discord.Net.Commands" Version="2.0.1" />
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="2.2.0" />
+    <PackageReference Include="Discord.Net.Commands" Version="3.4.1" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="6.0.0" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\Lavalink4NET.Discord_NET\Lavalink4NET.Discord_NET.csproj" />
-    <ProjectReference Include="..\..\src\Lavalink4NET.MemoryCache\Lavalink4NET.MemoryCache.csproj" />
+    <ProjectReference Include="G:\Lavalink4NET\src\Lavalink4NET.Discord_NET\Lavalink4NET.DiscordNet.csproj" />
+    <ProjectReference Include="G:\Lavalink4NET\src\Lavalink4NET.Logging.Microsoft\Lavalink4NET.Logging.Microsoft.csproj" />
+    <ProjectReference Include="G:\Lavalink4NET\src\Lavalink4NET.MemoryCache\Lavalink4NET.MemoryCache.csproj" />
   </ItemGroup>
 </Project>
